+++ src/ADAL.PCL.Desktop/DeviceAuthHelper.cs
@@ -29,18 +29,11 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
-using System.IdentityModel.Tokens;
-using System.IO;
-using System.Runtime.ConstrainedExecution;
-using System.Runtime.InteropServices;
-using System.Security;
 using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
-using System.Security.Permissions;
 using System.Text;
 using System.Threading.Tasks;
 using Microsoft.IdentityModel.Clients.ActiveDirectory.Native;
-using Microsoft.Win32.SafeHandles;
 
 namespace Microsoft.IdentityModel.Clients.ActiveDirectory
 {
@@ -58,7 +51,7 @@ namespace Microsoft.IdentityModel.Clients.ActiveDirectory
             X509Certificate2 certificate = FindCertificate(challengeData);
             DeviceAuthJWTResponse response = new DeviceAuthJWTResponse(challengeData["SubmitUrl"],
                 challengeData["nonce"], Convert.ToBase64String(certificate.GetRawCertData()));
-            CngKey key = GetCngPrivateKey(certificate);
+            CngKey key = CryptographyHelper.GetCngPrivateKey(certificate);
             byte[] sig = null;
             using (RSACng rsa = new RSACng(key))
             {
@@ -140,89 +133,6 @@ namespace Microsoft.IdentityModel.Clients.ActiveDirectory
             {
                 store.Close();
             }
-        }
-
-
-        /// <summary>
-        ///     <para>
-        ///         The GetCngPrivateKey method will return a <see cref="CngKey"/> representing the private
-        ///         key of an X.509 certificate which has its private key stored with NCrypt rather than with
-        ///         CAPI. If the key is not stored with NCrypt or if there is no private key available,
-        ///         GetCngPrivateKey returns null.
-        ///     </para>
-        ///     <para>
-        ///         The HasCngKey method can be used to test if the certificate does have its private key
-        ///         stored with NCrypt.
-        ///     </para>
-        ///     <para>
-        ///         The X509Certificate that is used to get the key must be kept alive for the lifetime of the
-        ///         CngKey that is returned - otherwise the handle may be cleaned up when the certificate is
-        ///         finalized.
-        ///     </para>
-        /// </summary>
-        /// <permission cref="SecurityPermission">The caller of this method must have SecurityPermission/UnmanagedCode.</permission>
-        [SecurityCritical]
-        [SuppressMessage("Microsoft.Security", "CA2122:DoNotIndirectlyExposeMethodsWithLinkDemands",
-            Justification = "Safe use of LinkDemand methods")]
-        public static CngKey GetCngPrivateKey(X509Certificate2 certificate)
-        {
-            using (SafeCertContextHandle certContext = GetCertificateContext(certificate))
-            using (SafeNCryptKeyHandle privateKeyHandle = X509Native.AcquireCngPrivateKey(certContext))
-            {
-                // We need to assert for full trust when opening the CNG key because
-                // CngKey.Open(SafeNCryptKeyHandle) does a full demand for full trust, and we want to allow
-                // access to a certificate's private key by anyone who has access to the certificate itself.
-                new PermissionSet(PermissionState.Unrestricted).Assert();
-                return CngKey.Open(privateKeyHandle, CngKeyHandleOpenOptions.None);
-            }
-        }
-
-        /// <summary>
-        ///     Get a <see cref="SafeCertContextHandle" /> for the X509 certificate.  The caller of this
-        ///     method owns the returned safe handle, and should dispose of it when they no longer need it. 
-        ///     This handle can be used independently of the lifetime of the original X509 certificate.
-        /// </summary>
-        /// <permission cref="SecurityPermission">
-        ///     The immediate caller must have SecurityPermission/UnmanagedCode to use this method
-        /// </permission>
-        [SecurityCritical]
-        [SuppressMessage("Microsoft.Reliability", "CA2004:RemoveCallsToGCKeepAlive",
-            Justification =
-                "This method is used to create the safe handle, and KeepAlive is needed to prevent racing the GC while doing so"
-            )]
-        public static SafeCertContextHandle GetCertificateContext(X509Certificate certificate)
-        {
-            SafeCertContextHandle certContext = X509Native.DuplicateCertContext(certificate.Handle);
-
-            // Make sure to keep the X509Certificate object alive until after its certificate context is
-            // duplicated, otherwise it could end up being closed out from underneath us before we get a
-            // chance to duplicate the handle.
-            GC.KeepAlive(certificate);
-
-            return certContext;
-        }
-    }
-
-
-    [SecurityCritical]
-    internal sealed class SafeCertContextHandle : SafeHandleZeroOrMinusOneIsInvalid
-    {
-        private SafeCertContextHandle()
-            : base(true)
-        {
-        }
-
-        [DllImport("crypt32.dll")]
-        [ReliabilityContract(Consistency.WillNotCorruptState, Cer.Success)]
-        [SuppressMessage("Microsoft.Design", "CA1060:MovePInvokesToNativeMethodsClass",
-            Justification = "SafeHandle release method")]
-        [SuppressUnmanagedCodeSecurity]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        private static extern bool CertFreeCertificateContext(IntPtr pCertContext);
-
-        protected override bool ReleaseHandle()
-        {
-            return CertFreeCertificateContext(handle);
         }
     }
 }
